% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize}
\alias{standardize}
\title{Standardize a numeric vector}
\usage{
standardize(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\description{
For a numeric vector, center the values such that the mean is zero, and scale such that the standard deviation is one.
}
